Decouple impl. from interface, excluding representation, digit vector, and some inline functions
